pyproject.toml
[tool.poetry]
name = "hybrid-trading-pipeline"
version = "0.1.0"
description = "Production-grade hybrid streaming + batch data pipeline for AI-driven crypto trading on Binance."
authors = ["Fah"]
license = "MIT"
readme = "README.md"
package-mode= false

[tool.poetry.dependencies]
python = "^3.11"
python-binance = "^1.0.19"
kafka-python = "^2.0.2"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0"
pandas = "^2.2"
xgboost = "^2.0"
scikit-learn = "^1.4"
streamlit = "^1.32"
python-dotenv = "^1.0"
structlog = "^24.1"
httpx = "^0.27"
websocket-client = "^1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
pytest-cov = "^4.1"
ruff = "^0.3"
mypy = "^1.9"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
omit = ["tests/*", ".venv/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
